requirements.txt
streamlit>=1.29.0,<1.40.0
Pillow>=9.0.0
streamlit-image-annotation>=0.0.5
